refactor: separate logic from api route
